pyproject.toml
[tool.poetry]
name = "2026-etpgrf-site"
version = "0.1.0"
description = ""
authors = ["erjemin <erjemin@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
django = "^6.0"
psycopg2-binary = "^2.9.11"
gunicorn = "^23.0.0"
python-dotenv = "^1.2.1"
# Зависимости для локальной библиотеки etpgrf
# etpgrf = {path = "./etpgrf", develop = true}
beautifulsoup4 = "^4.12"
lxml = "^5.1"
regex = "^2023.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
